+++ python/GafferArnoldTest/InteractiveArnoldRenderTest.py
@@ -129,7 +129,7 @@ class InteractiveArnoldRenderTest( GafferSceneTest.InteractiveRenderTest ) :
 
 		script["attributes"] = GafferArnold.ArnoldAttributes()
 		script["attributes"]["in"].setInput( script["meshType"]["out"] )
-		script["attributes"]["attributes"]["subdivIterations"]["enabled"].setValue( True )
+		script["attributes"]["attributes"]["ai:polymesh:subdiv_iterations"]["enabled"].setValue( True )
 
 		script["catalogue"] = GafferScene.Catalogue()
 
@@ -175,7 +175,7 @@ class InteractiveArnoldRenderTest( GafferSceneTest.InteractiveRenderTest ) :
 		# Now up the number of subdivision levels. The alpha coverage should
 		# increase as the shape tends towards the limit surface.
 
-		script["attributes"]["attributes"]["subdivIterations"]["value"].setValue( 4 )
+		script["attributes"]["attributes"]["ai:polymesh:subdiv_iterations"]["value"].setValue( 4 )
 		self.uiThreadCallHandler.waitFor( 1 )
 
 		self.assertAlmostEqual( script["imageStats"]["average"][3].getValue(), 0.424, delta = 0.001 )